Updated data type for repositories

+++ FRS/src/main/java/com/group/repository/FlightRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.group.model.Flight;
 
-public interface FlightRepository  extends CrudRepository<Flight, Long> {
+public interface FlightRepository  extends CrudRepository<Flight, Integer> {
