+++ FRS/src/main/java/com/group/repository/PassengerRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.group.model.Passenger;
 
-public interface PassengerRepository extends CrudRepository<Passenger, Long> {
+public interface PassengerRepository extends CrudRepository<Passenger, Integer> {
 
 }
